Move greetings to lib folder
